requirements.txt
absl-py==0.9.0
audioread==2.1.8
cachetools==4.0.0
cffi==1.14.0
cycler==0.10.0
decorator==4.4.1
google-auth==1.11.2
google-auth-oauthlib==0.4.1
grpcio==1.27.2
joblib==0.14.1
kiwisolver==1.1.0
librosa==0.7.2
llvmlite==0.31.0
Markdown==3.2.1
matplotlib==3.1.3
nnAudio==0.0.11
numba==0.48.0
oauthlib==3.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
pyparsing==2.4.6
python-dateutil==2.8.1
PyYAML==5.3
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.0
scikit-learn==0.22.1
six==1.14.0
SoundFile==0.10.3.post1
tensorboard==2.1.0
tqdm==4.43.0
Werkzeug==1.0.0
certifi==2019.11.28
chardet==3.0.4
dominate==2.4.0
idna==2.9
jsonpatch==1.25
jsonpointer==2.0
numpy==1.18.1
Pillow==7.0.0
protobuf==3.11.3
pyzmq==19.0.0
requests==2.23.0
scipy==1.4.1
tensorboardX==2.0
torch==1.4.0
torchfile==0.1.0
torchvision==0.5.0
tornado==6.0.4
urllib3==1.25.8
visdom==0.1.8.9
websocket-client==0.57.0
